Remove estimator, rename training loops main2 to main
