+++ setup.py
@@ -16,7 +16,7 @@ setup(
             'sp-train = lm.data:sp_train',
             'sp-encode = lm.data:sp_encode',
             'gpt-2-tf-train = lm.gpt_2_tf.train:main',
-            'gpt-2-tf2 = lm.gpt_2_tf2.main2:fire_main',
+            'gpt-2-tf2 = lm.gpt_2_tf2.main:fire_main',
         ],
     }
 )
